+++ src/hmc_link_buf.h
@@ -8,7 +8,7 @@ class hmc_notify;
 
 class hmc_link_buf {
 
-  unsigned bitoccupation;
+  float bitoccupation;
   unsigned bitoccupationmax;
   std::list< std::pair<char*,unsigned> > buf;
 
@@ -20,7 +20,7 @@ public:
 
   void adjust_size(unsigned bitsize);
 
-  bool reserve_space(unsigned packetleninbit);
+  bool reserve_space(float packetleninbit);
   void push_back_set_avail(char *packet, unsigned packetleninbit);
 
   char *front(unsigned *packetleninbit);
